rename rosstart to rosexecute
"rosstart" collides with "rosservice", "rossrv", and "rosstack"
with the resulting command input "rosst<tab>r<tab>".

"rosexecute" on the other hand only collides with "rosed" and
leads to "rosex<tab>".
I never noticed we are almost out of letters that complete in
"ros<letter><tab>". "rosx" would be the best choice, but wouldn't
read as well anymore.