Fixes for ROS-O, debian testing, and uncommon builds (#37)
* fix error

On ROS, the previous code evaluates to the string `if ! function ; then`.
That's a syntax error.

* the argcomplete command line script got renamed

from Ubuntu kinetic. They dropped the "3"
https://packages.ubuntu.com/search?suite=kinetic&arch=any&mode=exactfilename&searchon=contents&keywords=register-python-argcomplete

Support both script names.

* support non-/opt/ros installations of ROS

esp. the Debian ROS project and ROS-O source builds.

We really should not rely on /opt/ros/$ROS_DISTRO but look
for the executable we overlayed in PATH.

This touches ROS only because I can't test ROS2
in such an installation yet.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* async/pathlib version of get_overlayed_command

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: David V. Lu <davidvlu@gmail.com>